+++ src/js/actions/RolesAction.ts
@@ -1,0 +1,14 @@
+import GenericCRUDActions from "./GenericCRUDActions";
+import { instance as rolesAPI } from "../../api/rolesAPI";
+
+export default class RolesActions extends GenericCRUDActions {
+    
+    constructor() {
+
+        super("roles", rolesAPI);
+        
+    }
+
+}
+
+export const actions = new RolesActions();