+++ src/js/reducers/index.js
@@ -1,10 +1,12 @@
 import { combineReducers } from 'redux';
 import ping from "./app/ping";
 import login from "./app/login";
+import roles from "./app/roles";
 
 export default combineReducers(
   {
     ping,
     login,
+    roles,
   }
 );